+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Discljord follows semantic versioning.
 
 ## [Unreleased]
 
+ - Support for stream-like data in `discljord.messaging/create-message!`
+
 ## [0.2.8] - 2020-04-13
 ### Added
  - Support for audit log reasons on all requests made with `discljord.messaging`